refactor(contract): 이벤트 패키지 변경
