+++ contract-service/src/main/java/com/example/contractservice/contract/service/batch/writer/ContractInProgressWriter.java
@@ -2,7 +2,7 @@ package com.example.contractservice.contract.service.batch.writer;
 
 import com.example.contractservice.contract.common.ContractStatus;
 import com.example.contractservice.contract.entity.ContractEntity;
-import com.example.contractservice.contract.event.dto.ContractEvent;
+import com.example.contractservice.contract.service.event.dto.ContractEvent;
 import com.example.contractservice.contract.repository.ContractRepository;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
