+++ contract-service/src/main/java/com/example/contractservice/contract/service/event/ContractEventProducer.java
@@ -1,6 +1,6 @@
-package com.example.contractservice.contract.event;
+package com.example.contractservice.contract.service.event;
 
-import com.example.contractservice.contract.event.dto.ContractEvent;
+import com.example.contractservice.contract.service.event.dto.ContractEvent;
 import java.util.concurrent.CompletableFuture;
 import org.springframework.kafka.support.SendResult;
